refactor: split runtime quota settings models
